backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["SmokierLemur51 <ldl6147@gmail.com>"]
readme = "README.md"
packages = [{include = "backend", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
quart = "^0.19.6"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.1"
quart-db = {extras = ["postgresql"], version = "^0.8.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "backend:run"
